PFS-274: Fixes int conversions by safely mapping to the expected database bit size. Uses scanner functions to map fields instead of manually specifying the validity.

+++ finance-api/internal/store/queries/invoices.sql
@@ -95,7 +95,7 @@ WHERE i.id = $1;
 -- name: GetInvoiceBalancesForFeeReductionRange :many
 SELECT i.id,
        i.amount,
-       COALESCE(general_fee.amount, 0)               general_supervision_fee,
+       COALESCE(general_fee.amount, 0)::INT          general_supervision_fee,
        i.amount - COALESCE(transactions.received, 0) outstanding,
        i.feetype
 FROM invoice i
